HTBHFB-115 Rename domain package to entity
