+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/entity/Claimant.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.domain;
+package uk.gov.dhsc.htbhf.claimant.entity;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
