+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/repository/ClaimantRepository.java
@@ -1,7 +1,7 @@
 package uk.gov.dhsc.htbhf.claimant.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import uk.gov.dhsc.htbhf.claimant.domain.Claimant;
+import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 
 import java.util.List;
