+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/domain/ClaimantTest.java
@@ -2,6 +2,7 @@ package uk.gov.dhsc.htbhf.claimant.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 import uk.gov.dhsc.htbhf.claimant.testsupport.ClaimantTestDataFactory;
 
 import java.util.Set;
