+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/repository/ClaimantRepositoryTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.dhsc.htbhf.claimant.domain.Claimant;
+import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 
 import java.util.List;
 import javax.transaction.Transactional;
